sonarqube and database related changes

+++ src/main/java/assignment/individualtrack/domain/Game/StartGameRequest.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 @Builder
 public class StartGameRequest {
-    private int playerID;
+    private Long playerID;
 }
